% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.VegLog.R
\name{as.VegLog}
\alias{as.VegLog}
\title{Coerce existing data frame to VegLog object.}
\usage{
as.VegLog(x)
}
\arguments{
\item{x}{data frame containing standard vegetation columns}
}
\value{
VegLog data containing slots for site data and plant data
}
\description{
Coerce existing data frame to VegLog object.
}
\examples{
obssites <- vegnasis::obs
obstaxa <- vegnasis::obsspp
veg=clean.veg.log(obssites, obstaxa)
manyplots = as.VegLog(veg)
}
