% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.VegLog.R
\name{c,VegLog-method}
\alias{c,VegLog-method}
\title{Merge several VegPlot objects}
\usage{
\S4method{c}{VegLog}(x, ...)
}
\arguments{
\item{VegLog}{one or more VegLog objects}
}
\value{
new larger VegLog object
}
\description{
Merge several VegPlot objects
}
\examples{
obssites <- vegnasis::obs
obstaxa <- vegnasis::obsspp
veg=clean.veg.log(obssites, obstaxa)
veg1 = as.VegLog(veg[1:10,])
veg2 = as.VegLog(veg[50:60,])
veg3 = c(veg1,veg2)
}
