% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.agg.R
\name{carea}
\alias{carea}
\title{Get total crown area index from aggregate canopy area.}
\usage{
carea(k, avoid = T)
}
\arguments{
\item{k}{Aggregate canopy cover}

\item{avoid}{TRUE: Crowns compete for light and avoid overlapping; FALSE: Crowns are completely random.}
}
\value{
summed crown area (crown area index), which may exceed 100%.
}
\description{
Get total crown area index from aggregate canopy area.
}
