% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.veg.R
\name{clean.veg}
\alias{clean.veg}
\title{Clean vegetation data imported from NASIS}
\usage{
clean.veg(x, type = TRUE, hts = FALSE)
}
\arguments{
\item{x}{Species composition data frame derived from a soilDB NASIS import function.}

\item{type}{Should missing plant types (habits) be filled (TRUE/FALSE)?}

\item{hts}{Should missing plant heights be filled to allow assignment to strata (TRUE/FALSE)?}
}
\value{
Data frame containing a standardized set of quanitative vegetation parameters.
}
\description{
This function takes veg plot imported from NASIS and consolidates the redundant cover and height fields.
}
\examples{
veg.raw <- soilDB::get_vegplot_species_from_NASIS_db()
veg <- clean.veg(veg.raw)

}
