% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plants.R
\name{colormixer}
\alias{colormixer}
\title{Color mixer}
\usage{
colormixer(colorname, mixcolor, p)
}
\arguments{
\item{colorname}{Name or hex code for main color of object (can include alpha for transparency).}

\item{mixcolor}{Color used for fading into (maintaining the same alpha as main color).}

\item{p}{proportion of fading into the mix color.}
}
\value{
Faded color.
}
\description{
Color mixer
}
\examples{
crowncolor='darkgreen'; fadecolor = "#D9F2FF"
colormixer(crowncolor, fadecolor, 0.5)
}
