% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.agg.R
\name{cover.agg}
\alias{cover.agg}
\title{Aggregate canopy cover}
\usage{
cover.agg(x, avoid = FALSE)
}
\arguments{
\item{x}{Cover of individual taxa or strata to aggregate.}

\item{avoid}{TRUE: Crowns compete for light and avoid overlapping; FALSE: Crowns are completely random.}
}
\value{
Aggregate cover
}
\description{
This function aggregates cover among taxa of a given stratum or guild, assuming random crown distributions, and where totals cannot not exceeding 100 percent. This function is called within many vegetation summary functions. Essentially the same as a method suggested by an equation in Appendex K of the US National Vegetation Classification (USNVC, 2008). The crown avoidance option gives more priority for stems located in opening rather than in the shade of another crown.
}
\examples{
x <- c(5,5,10,20)
cover.agg(x, avoid=TRUE)
cover.agg(x, avoid=FALSE)
}
