% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTaxon.R
\name{extractTaxon}
\alias{extractTaxon}
\title{Extract Taxon}
\usage{
extractTaxon(rawnames, report = "taxon")
}
\arguments{
\item{rawnames}{Vector of taxonomic names with or without authors and infrataxa.}

\item{report}{What element of taxon name to report ("taxon", "genus", "binomial", "author").}
}
\value{
Name of taxon or author depending on report parameter ("taxon" = full taxon without author; "binomial" = genus and specific epithet; "author" = author of of lowest infrataxon for which author name is provided.)
}
\description{
This function takes a full taxonomic name complete with authors and infrataxa (variety and subspecies and forms) and parses out the different elements. Hybrids indicated by the letter 'X' are converted to the multiply '×' symbol and spaces are removed so that hybrid species can be treated with the same number of word elements as non hybrids.
}
\examples{
rawnames = c('Arnica angustifolia Vahl subsp. tomentosa (Macoun) G.W. Douglas & G. Ruyle-Douglas', 'Abies ×shastensis (Lemmon) Lemmon', 'Abies X shastensis (Lemmon) Lemmon','Abies balsamea ssp. lasiocarpa', 'Abies balsamea (L.) Mill. var. fallax (Engelm.) B. Boivin','Agastache pallidiflora (A. Heller) Rydb. ssp. pallidiflora var. greenei (Briq.) R.W. Sanders')
extractTaxon(rawnames,'taxon')
extractTaxon(rawnames,'binomial')
extractTaxon(rawnames,'author')

}
