% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.agg.R
\name{findcw}
\alias{findcw}
\title{Find crown width from canopy cover and stem density.}
\usage{
findcw(k, st, a = 1, avoid = T)
}
\arguments{
\item{k}{canopy cover}

\item{st}{number of stems per unit area}

\item{a}{area (hectares)}

\item{avoid}{TRUE: Crowns compete for light and avoid overlapping; FALSE: Crowns are completely random.}
}
\value{
crown width (meters)
}
\description{
Find crown width from canopy cover and stem density.
}
