% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.strata.R
\name{flat.summary}
\alias{flat.summary}
\title{Flatten a summary of multiple relevé vegetation plots.}
\usage{
flat.summary(veg.summ)
}
\arguments{
\item{veg.summ}{Data frame results from summary.ESIS() function.}
}
\value{
Data frame with the multiple strata as row coverted to columns and displaying each taxon in a single row.
}
\description{
Flatten a summary of multiple relevé vegetation plots.
}
\examples{
veg <- vegnasis::nasis.veg |> clean.veg() #Get example data.
veg <- subset(veg, plot \%in\% unique(veg$plot)[1:5]) |> fill.hts.df() #Get arbitrary subset of example data.
veg.summ <-  summary.ESIS(veg) #Mutliple relevé summary.
veg.flat <- flat.summary(veg.summ) #Flatten multiple strata rows to columns.
# write.csv(veg.flat, 'veg.flat.csv', row.names = FALSE) #Save as spreadsheet to share.
}
