% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wetness.R
\name{get.obl}
\alias{get.obl}
\title{Get obligate/fac-wet indicator status}
\usage{
get.obl(x, region = "NCNE")
}
\arguments{
\item{x}{Species composition data frame cleaned with clean.veg() function.}

\item{region}{US Army Corps Region where indicator status is prioritized.}
}
\value{
wetness value
}
\description{
This function gets wetness status using only obligate and faculative-wet species (obligate having twice the weights as fac-wet), and rates all faculative and upland species as zero.\n
Army corps regions: 'AGCP' = Atlantic and Gulf Coastal Plain; 'AW' = Arid West; 'CB' = Caribbean; 'EMP' = Eastern Mountains and Piedmont; 'GP' = Great Plains; 'HI' = HI; 'MW' = Midwest; 'NCNE' = Northcentral & Northeast; 'AK' = Alaska; 'WMVC' = Western Mountains, Valleys, and Coast.
}
\examples{
data("nasis.veg")
veg <- nasis.veg |> clean.veg()
wetland <- veg |> get.obl()
}
