% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wetness.R
\name{get.upl}
\alias{get.upl}
\title{Get upland/fac-up indicator status}
\usage{
get.upl(x, region = "NCNE")
}
\arguments{
\item{x}{Species composition data frame cleaned with clean.veg() function.}

\item{region}{US Army Corps Region where indicator status is prioritized.}
}
\value{
dryness value
}
\description{
This function gets upland status using only upland and faculative-upland species (upland having twice the weights as fac-up), and rates all faculative and wetland species as zero.\n
Army corps regions: 'AGCP' = Atlantic and Gulf Coastal Plain; 'AW' = Arid West; 'CB' = Caribbean; 'EMP' = Eastern Mountains and Piedmont; 'GP' = Great Plains; 'HI' = HI; 'MW' = Midwest; 'NCNE' = Northcentral & Northeast; 'AK' = Alaska; 'WMVC' = Western Mountains, Valleys, and Coast.
}
\examples{
data("nasis.veg")
veg <- nasis.veg |> clean.veg()
wetland <- veg |> get.upl()
}
