% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{hull.buffer}
\alias{hull.buffer}
\title{Buffer around or within convex hull.}
\usage{
hull.buffer(x, y, s, b)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}

\item{s}{vector of integers determining the ordering of coordinates.}

\item{b}{Buffer thickness (use negative number for internal buffer)}
}
\value{
Data frame of xy coordinates of densified points following the border of hull.
}
\description{
Buffer around or within convex hull.
}
