% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdominant.R
\name{isdominant}
\alias{isdominant}
\title{Determine dominant species}
\usage{
isdominant(veg = NULL, plot, taxon, cover, ht.max, type, append = FALSE)
}
\arguments{
\item{veg}{Standardized vegetation data frame}

\item{plot}{Vector of plot ids (in lieu of data frame)}

\item{taxon}{Vector of taxon names (in lieu of data frame)}

\item{cover}{Vector of cover (in lieu of data frame)}

\item{ht.max}{Vector of maximum plant height (in lieu of data frame)}

\item{type}{Vector of functional groups (in lieu of data frame)}

\item{append}{Should add dominant designation as column (TRUE), or show data summarize data frame to 3 strata from which dominance was determined (FALSE)}
}
\value{
Data frame identifying dominant species as 1, or 0 for non-dominant species in a column called "isdom".
}
\description{
Determine dominant species
}
\examples{
#create demo data set
taxon=c('Tree1', 'Tree2','Tree2', 'Tree3', 'Shrub1', 'Shrub2', 'Shrub3','Tree1', 'Herb1', 'Herb2', 'Herb3')
cover=c(10,50,2,1,19,5,14,1,22,15,2)
ht.max=c(20,15,6,10,3,2,0.3,2,0.2,0.5,0.1)
type = c('tree','tree','tree','tree','shrub/vine','shrub/vine','shrub/vine','tree','forb','forb','grass/grasslike')
plot = 'plot1'
df <- data.frame(plot=plot,taxon=taxon,cover=cover,ht.max=ht.max,type=type)
#input as data frame, appending
isdominant(df, append = TRUE)
#input as data frame, summarizing strata
isdominant(df, append = FALSE)
#input as vectors
isdominant(plot=plot,taxon=taxon,cover=cover,ht.max=ht.max,type=type)
}
