% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{makeCrownShape}
\alias{makeCrownShape}
\title{Make Tree Crown Outline}
\usage{
makeCrownShape(
  ht.max = 5,
  ht.min = 1,
  crwd = 2,
  dbh = 0.3,
  tip = 0.01,
  crshape = c("pyramid", "dome", "round", "column"),
  n = 5,
  bu = 0.8,
  bl = 0,
  opposite = FALSE
)
}
\arguments{
\item{ht.max}{Height of the tree.}

\item{ht.min}{Height of the lowest branches.}

\item{crwd}{Crown width.}

\item{dbh}{Stem base diameter.}

\item{tip}{Thickness of stem tip.}

\item{crshape}{Basic crown shape (pyramid, dome, round, or column).}

\item{n}{Number of branches.}

\item{bu}{Relative position (0-1) of highest branch attachment to tree relative to the top and bottom of the crown.}

\item{bl}{Relative position (0-1) of lowest branch attachment to tree relative to the top and bottom of the crown.}

\item{opposite}{If true the all branches will be placed opposite each other on stem. If false, the branches will alternate the right and left side of stem.}
}
\value{
A data frame with information for attaching branches to a stem.
}
\description{
Make Tree Crown Outline
}
\examples{
#fractal tree with bend ---
library(ggplot2)
n=3
lth=4
wth = 0.5
sc = 0.6
stem <-  makeStem(lth,wth,0.1*wth,20)
stem <-  stem |> mutate(x = x + -0.2*cos(y/lth*2*pi), center = center + -0.2*cos(y/lth*2*pi))
stem <-  skewStem(stem,-0.2,.5,3)
branch <-  stem |> mutate(x=x*sc,y=y*sc,center=center*sc)
branch2 <-  stem |> mutate(x=x*sc*0.7,y=y*sc*0.7,center=center*sc*0.7)
branch3 <-  stem |> mutate(x=x*sc*0.3,y=y*sc*0.3,center=center*sc*0.3)
for(i in 1:n){
  tree <- attachBranch(stem, branch, -50, 2)
  branch <- branch2
  tree <- attachBranch(tree, branch, 50, 3)
  branch <- branch3
  #stem=tree
  tree <- attachBranch(tree, branch, -5, 4)
  branch <- tree |> mutate(x=x*sc,y=y*sc,center=center*sc)
  branch2 <- tree |> mutate(x=x*sc*.7,y=y*sc*.7,center=center*sc*.7)
  branch3 <- tree |> mutate(x=x*sc*.3,y=y*sc*.3,center=center*sc*.3)
}

crown <- tree |> subset(grepl('tip',type) & !type \%in\% 'tip')
ggplot()+
  geom_polygon(data=tree, aes(x=x, y=y), color='brown',fill='#99500090')+
  # geom_point(data=stem, aes(x=x, y=y), color='red')+
  geom_polygon(data=crown, aes(x=x, y=y), color='green',fill='#00990090')+
  # geom_point(data=tree, aes(x=x, y=y), color='green')+
  coord_fixed()

  #------conifer

crshape = c('pyramid','dome','round','column')

crshape = c('pyramid','column')
shapes <- makeCrownShape(ht.max=10, ht.min=3, crwd=3, dbh=0.5, n=7, bu=1, bl=0.2, crshape=crshape,
                         opposite = T)
shapes <- subset(shapes, !(a > 175 | a < -175 | a == 0)  & l> 0.3)
stem <-  makeStem(10,0.5,0.01,10)
for(i in 1:nrow(shapes)){#i=1
  branch <- makeStem(shapes$l[i], shapes$d[i]*.5,0.01,10)
  bpos <- max(branch$y)
  branch2 <- branch |> mutate(x=x*0.3,y=y*0.3)
  branch3 <- branch |> mutate(x=x*0.2,y=y*0.2)

  branch <- skewStem(branch, amp=ifelse(shapes$a[i] >= 0,-0.07*(shapes$s[i]*-1+1),0.07*(shapes$s[i]*-1+1)),
                     phase=0, waves=1)
  branchA <- attachBranch(branch, branch2, ifelse(shapes$a[i] >= 0,30,-30), bpos*0.3)
  branchA <- attachBranch(branchA, branch3, ifelse(shapes$a[i] >= 0,-30,30), bpos*0.7)


  stem <- attachBranch(stem, branchA, shapes$a[i], shapes$by[i])
}
crown <- stem |> subset(grepl('tip',type) | type \%in\% 'bbase')


ggplot()+
  geom_polygon(data=stem, aes(x=x, y=y), color='brown',fill='#99500090')+
  # geom_point(data=stem, aes(x=x, y=y), color='red')+
  geom_polygon(data=crown, aes(x=x, y=y), color='green',fill='#00990090')+
  # geom_point(data=tree, aes(x=x, y=y), color='green')+
  coord_fixed()

  #------cottonwood

  crshape = c('pyramid','dome','round','column')

  crshape = c('dome')
  shapes <- makeCrownShape(ht.max=10, ht.min=5, crwd=6, dbh=0.5, n=3, bu=0.5, bl=0, crshape=crshape,
  opposite = F)
  shapes <- subset(shapes, !(a > 175 | a < -175 | a==0)  & l> 0.3)
  stem <-  makeStem(10,0.5,0.01,25)
  stem <- skewStem(stem, amp=-0,
                   phase=0, waves=0.5)

  for(i in 1:nrow(shapes)){#i=1
    branch <- makeStem(shapes$l[i], shapes$d[i]*.5,0.01,15)
    bpos <- max(branch$y)
    branch2 <- branch |> mutate(x=x*0.3,y=y*0.3)
    branch3 <- branch |> mutate(x=x*0.2,y=y*0.2)

    branch <- skewStem(branch, amp=ifelse(shapes$a[i] >= 0,-0.2,0.2),
                       phase=0, waves=1)
    branchA <- attachBranch(branch, branch2, ifelse(shapes$a[i] >= 0,30,-30), bpos*0.5)
    branchA <- attachBranch(branchA, branch3, ifelse(shapes$a[i] >= 0,-30,30), bpos*0.6)


  stem <- attachBranch(stem, branchA, shapes$a[i], shapes$by[i], tht = shapes$ty[i], tx = shapes$tx[i])
  }

  branchB <- stem |> mutate(x=x*1,y=y) |> skewStem(amp=0.3)

  stem2 <-  makeStem(5,0.5,0.3,30)
  tree <- attachBranch(stem2, branchB, bht=4.5, tht = 11, tx=-4)
  tree <- attachBranch(tree, branchB, bht=5, tht = 11, tx=3)

  crown <- tree |> subset(grepl('tip',type))
  crown2 <- crown[chull(x=crown$x, y=crown$y),]

  ggplot()+
   geom_polygon(data=tree, aes(x=x, y=y), color='brown',fill='#99500050')+
   # geom_point(data=stem, aes(x=x, y=y), color='red')+
   geom_polygon(data=crown2, aes(x=x, y=y), color='green',fill='#00990050')+
    # geom_point(data=tree, aes(x=x, y=y), color='green')+
    coord_fixed()

}
