% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{makeStem}
\alias{makeStem}
\title{Make Stem}
\usage{
makeStem(lth, wth, tip = 0.01, inc = 10)
}
\arguments{
\item{lth}{Length of stem.}

\item{wth}{Width of stem at base.}

\item{tip}{Width of stem at tip.}

\item{inc}{Number of vertices increments to include (how many dots per side).}
}
\value{
Simple stem data frame with xy coordinates for plotting.
}
\description{
Make Stem
}
