% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{minXY}
\alias{minXY}
\title{Minimum width of convex hull}
\usage{
minXY(x, y)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}
}
\value{
Width of convex hull.
}
\description{
Minimum width of convex hull
}
