% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.agg.R
\name{nstem}
\alias{nstem}
\title{Estimate number of stems from crown width and canopy cover}
\usage{
nstem(k, cw, a = 1, avoid = T)
}
\arguments{
\item{k}{canopy cover}

\item{cw}{crown width (m)}

\item{a}{area (default 1 hectare)}

\item{avoid}{TRUE: Crowns compete for light and avoid overlapping; FALSE: Crowns are completely random.}
}
\description{
Estimate number of stems from crown width and canopy cover
}
