% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{renumber}
\alias{renumber}
\title{Renumber a set of numbers}
\usage{
renumber(x)
}
\arguments{
\item{x}{vector of numbers, including no data.}
}
\value{
vector of numbers converted to sequential integers, ignoring but maintaining no data. Used inside other functions which rearrange vertices for plotting polygons.
}
\description{
Renumber a set of numbers
}
