% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_plants.R
\name{setplants.overhead}
\alias{setplants.overhead}
\title{Assign plant shapes to stem locations for overhead.}
\usage{
setplants.overhead(strats, stand)
}
\arguments{
\item{strats}{Stratum data frame.}

\item{stand}{Stand data frame.}
}
\value{
Data frame with coordinates to plot plant shapes arranged in 2 dimensions for a vegetation overhead diagram.
}
\description{
Assign plant shapes to stem locations for overhead.
}
