% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_plants.R
\name{setstand}
\alias{setstand}
\title{Set stems in stand}
\usage{
setstand(strats, plength = 50, pwidth = 20)
}
\arguments{
\item{strats}{Stratum data frame (as generated by setstrats() function).}

\item{plength}{Length of plot diagram in meters.}

\item{pwidth}{Width of plot diagram in meters.}
}
\value{
Stand data frame with stem locations.
}
\description{
Function semi-randomly determines location of stems, avoiding shade.
}
