% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_plants.R
\name{setstrats}
\alias{setstrats}
\title{Assign strata to plant data.}
\usage{
setstrats(veg, plength = 50, pwidth = 20)
}
\arguments{
\item{veg}{Vegetation plot data frame.}

\item{plength}{Length of plot diagram in meters.}

\item{pwidth}{Width of plot diagram in meters.}
}
\description{
Assign strata to plant data.
}
