% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{skewStem}
\alias{skewStem}
\title{Skew Stem}
\usage{
skewStem(stem, amp = 0.2, phase = 0, waves = 1)
}
\arguments{
\item{stem}{Name of stem object.}

\item{amp}{Amplitude of bend.}

\item{phase}{Shift in the position of the bend in the stem.}

\item{waves}{Number of bends in the stem.}
}
\value{
Stem object with bends introduced to make stem look twisted.
}
\description{
Skew Stem
}
