% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_plants.R
\name{splitveg}
\alias{splitveg}
\title{Split vegetation along gradient}
\usage{
splitveg(..., plength = 50, pwidth = 20, px)
}
\arguments{
\item{...}{Two or more vegetation data frames. Will iterate if more breaks than vegetation types.}

\item{plength}{Length of plot diagram in meters.}

\item{pwidth}{Width of plot diagram in meters.}

\item{px}{Vector of vegetation break points along the lengths as defined in plength.}
}
\value{
Plants data frame with vegetation mosaic ready for plotting as a profile diagram.
}
\description{
This function allows for multiple vegetation structures or compositions along a gradient set at user define positions.
}
