% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.agg.R
\name{tocov}
\alias{tocov}
\title{Aggregate canopy cover from summed crown area}
\usage{
tocov(kk, avoid = T)
}
\arguments{
\item{kk}{Total crown area}

\item{avoid}{TRUE: Crowns compete for light and avoid overlapping; FALSE: Crowns are completely random.}
}
\value{
Aggregate canopy cover
}
\description{
Aggregate canopy cover from summed crown area
}
