% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternateplot.R
\name{totriangle}
\alias{totriangle}
\title{Transform to ternary xy coordinates}
\usage{
totriangle(a = NA, b = NA)
}
\arguments{
\item{a}{Composition from bottom to top corner of triangle.}

\item{b}{Composition from upper left to lower right corner of triangle.}
}
\value{
xy coordinates for plotting onto ternary graph.
}
\description{
This function transforms three way composition to ternary xy coordinates for ploting onto a tiangle. The inputs need only the first two axis (a is 100% at top corner, b is 100% at lower right corner). The third axis is assumed to be the difference from the first two axes and 100%. If the three axis do not add up to 100%, they must be balanced with the toabc() function. Tree and shrub data should be transformed using the tovtern() function.
}
