% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternateplot.R
\name{tovtern}
\alias{tovtern}
\title{Transform tree and shrub composition data to ternary xy coordinates.}
\usage{
tovtern(veg)
}
\arguments{
\item{veg}{A vegetation plot data frame with key information on habit and plant hight by species.}
}
\value{
Data frame rendering the xy coordinates to plot on ternary graph.
}
\description{
Transforms tree and shrub composition data to ternary xy coordinates. Aggregates the total tree cover, and the total woody plant cover, then derives the shrub cover by subtracting tree cover from shrub cover (shrubs only counted as cover in the portion where trees are not overhead.
}
