% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_profile_plot.overhead.R
\name{veg_overhead_plot}
\alias{veg_overhead_plot}
\title{Draw plant overhead diagram of vegetation}
\usage{
veg_overhead_plot(
  plants,
  units = "m",
  gridalpha = 0.3,
  groundcolor = "#808066",
  xlim = c(0, 50),
  ylim = c(15, 35),
  xticks = 5
)
}
\arguments{
\item{plants}{Processed strata with overhead plant templates.}

\item{units}{Sets the units of measure. Default is meters ("m"). Feet is an alternative ("ft").}

\item{gridalpha}{A number from 0 to 1 to vary the visibility (opacity) of the background scale grid. Use 0 for no grid.}

\item{groundcolor}{Sets the color of the bare ground.}

\item{xlim}{A vector setting the maximum and minimum limits of the x-axis of the graph (meters only). Default is 0-50 m.}

\item{ylim}{A vector setting the maximum and minimum limits of the y-axis of the graph (meters only). Default is 15-35 m.}

\item{xticks}{The interval for major grid lines along the x-axis, expressed in prescribed units of measure (default is 5 m).}
}
\value{
ggplot object featuring vegetation overhead diagram.
}
\description{
Draw plant overhead diagram of vegetation
}
\examples{
veg.raw <-  vegnasis::nasis.veg
veg <- clean.veg(veg.raw)
veg <- subset(veg,  grepl('2022MI165021.P',plot))
strats <- setstrats(veg, plength = 50, pwidth=20)
stand <- setstand(strats, plength = 50, pwidth=20)
plants <- setplants.overhead(strats, stand)
veg_overhead_plot(plants)
}
