% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternateplot.R
\name{vtern}
\alias{vtern}
\title{Make ternary graph for vegetation structure}
\usage{
vtern(axislabs = c("Tree", "Shrub", "Herb"), background = F, grid = T, num = T)
}
\arguments{
\item{axislabs}{Vector of three axis labels for left, right, and bottom sides, respectively.}

\item{background}{Should color background with predefined structural categories be displayed?}

\item{grid}{Should grid lines for every 10 percent be displayed?}

\item{num}{Should grid values along axes be labeled?}
}
\value{
ggplot object onto which additional 3-way compositions may be graphed, assuming that their composition is transformed to triangle coordinates.
}
\description{
Make ternary graph for vegetation structure
}
\examples{
#load sample data
siteass <- vegnasis::siteass20250414
sites <- vegnasis::sites20250414
veg.raw <- vegnasis::veg.raw20250414
vegplot <- vegnasis::vegplot20250414
#select example plots
thesesites <- subset(siteass, usiteassocid \%in\% 'DSP-F094AB019MI-2024')
sites <- subset(sites, usiteid \%in\% thesesites$usiteid)
#get related information about the plots
landuse <- data.frame(siteobsiid=sites$siteobsiid, landuse = sites$commphasename)
landuse <- landuse |> left_join(data.frame(siteobsiid=vegplot$siteobsiid, plot = vegplot$vegplotid)) |> subset(!is.na(plot) & !is.na(landuse))
#fill in missing height and growth habit data
veg <- clean.veg(veg.raw)|> subset(!is.na(taxon)) |> fill.type.df() |> fill.hts.df()
#Aggregate total tree and shrub cover then transform to xy coordinates on triangle (shrub values are subtracted from aggregate woody plant totals so that values add up to 100\%).
veg.tern <- tovtern(veg) |> inner_join(landuse)
#plot onto triangle
vtern(background = T)+
geom_point(data=veg.tern, aes(x=x, y=y, shape = landuse), size=3)+
geom_text(data=veg.tern, aes(x=x, y=y, label=label), vjust=-1, hjust=1, size=2)+
scale_shape_manual(values=c(17, 16, 8))+
theme(legend.text = element_text(size = 8),legend.position='left')
}
