% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{attachBranch}
\alias{attachBranch}
\title{Attach branch to stem}
\usage{
attachBranch(stem, branch, angle = 90, bht, tht = NA, tx = NA)
}
\arguments{
\item{stem}{Name of the primary stem object.}

\item{branch}{Name of secondary stem object to be used as a branch.}

\item{angle}{Angle of branch from the vertical in degrees.}

\item{bht}{Height of branch attachment point on stem.}

\item{tht}{Optional total height of tree instead of specifying angle.}

\item{tx}{}
}
\value{
Stem object with attached branch.
}
\description{
Attach branch to stem
}
