% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrees.R
\name{cavhull}
\alias{cavhull}
\title{Concave Hull}
\usage{
cavhull(x, y, concave = TRUE)
}
\arguments{
\item{x}{Vector of x coordinates.}

\item{y}{Vector of x coordinates.}

\item{concave}{Return concave if true, convex if false.}
}
\value{
Vector of indices which determine the ordering of the appropriate points to create an outline around a set of xy points.
}
\description{
Concave Hull
}
